+++ libs/core/src/lib/utils/drag-and-drop/drag-and-drop.module.ts
@@ -3,10 +3,12 @@ import { CommonModule } from '@angular/common';
 import { DndListDirective } from './dnd-list/dnd-list.directive';
 import { DndItemDirective } from './dnd-item/dnd-item.directive';
 import { DragDropModule } from '@angular/cdk/drag-drop';
+import { DndKeyboardItemDirective } from './dnd-keyboard-item/dnd-keyboard-item.directive';
+import { DndKeyboardGroupDirective } from './dnd-keyboard-group/dnd-keyboard-group.directive';
 
 @NgModule({
     imports: [CommonModule, DragDropModule],
-    exports: [DndItemDirective, DndListDirective],
-    declarations: [DndListDirective, DndItemDirective]
+    exports: [DndItemDirective, DndListDirective, DndKeyboardGroupDirective, DndKeyboardItemDirective],
+    declarations: [DndItemDirective, DndListDirective, DndKeyboardGroupDirective, DndKeyboardItemDirective]
 })
 export class DragAndDropModule {}